+++ src/rmhmc.cpp
@@ -28,7 +28,7 @@ mcmc::rmhmc_int(const arma::vec& initial_vals, arma::mat& draws_out, std::functi
     bool success = false;
 
     const double BIG_NEG_VAL = MCMC_BIG_NEG_VAL;
-    const int n_vals = initial_vals.n_elem;
+    const size_t n_vals = initial_vals.n_elem;
 
     //
     // settings
@@ -39,12 +39,12 @@ mcmc::rmhmc_int(const arma::vec& initial_vals, arma::mat& draws_out, std::functi
         settings = *settings_inp;
     }
 
-    const int n_draws_keep   = settings.hmc_n_draws;
-    const int n_draws_burnin = settings.hmc_n_burnin;
+    const size_t n_draws_keep   = settings.hmc_n_draws;
+    const size_t n_draws_burnin = settings.hmc_n_burnin;
 
     const double step_size = settings.hmc_step_size;
-    const int n_leap_steps = settings.hmc_leap_steps;
-    const int n_fp_steps = settings.rmhmc_fp_steps;
+    const size_t n_leap_steps = settings.hmc_leap_steps;
+    const size_t n_fp_steps = settings.rmhmc_fp_steps;
 
     const bool vals_bound = settings.vals_bound;
     
@@ -78,7 +78,7 @@ mcmc::rmhmc_int(const arma::vec& initial_vals, arma::mat& draws_out, std::functi
     = [target_log_kernel, vals_bound, bounds_type, lower_bounds, upper_bounds] (const arma::vec& pos_inp, const arma::vec& mntm_inp, void* target_data, const double step_size, const arma::mat& inv_tensor_mat, const arma::cube& tensor_deriv, arma::mat* jacob_matrix_out) \
     -> arma::vec 
     {
-        const int n_vals = pos_inp.n_elem;
+        const size_t n_vals = pos_inp.n_elem;
         arma::vec grad_obj(n_vals);
 
         if (vals_bound)
